Cert and puppet should be in base ami
This is a very oppinionated change. Old version
installed puppet upon configuration and certs was created and stored
on the aminator node before being moved to the chroot.

This change moves that resposibility to the base ami, which is now
assumed to contain puppet as well as the neccesary certs.